Project first commit
Posting the Table.js project, which allows a developer to quickly create
spreadsheet style tables and use them to store data.

Included in this Table class are some basic functions to add rows,
columns, get the size of the table, and of course instantiation. See the
readme file and the 'contribute' section to see features that will be
implemented.