Table tabbing, project renamed
Project was renamed from Table.js to Spreadsheet.js to be more clear on
what the library does. Also has the added benefit of avoiding ambiguity
between the Table object, and the html table it contained.

User can now tab through cells, which will select the next cell after
the one which is currently focused. It overflows to the next row, if
necessary.

Focusing on a cell now selects all of its text for easier editing.